Refactored server to use an event-based programming model via broker queue subscription rather than a busy while loop
